requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
faiss-cpu
sentence-transformers
openai==0.28

